alpha_v2

+++ modules/ads/ad_delete.php
@@ -8,16 +8,24 @@
    $listingService = new ListingService();
    
    $addInfo = $adService->single([
-    'where' => [
-        'ads.recno' => $addId
-    ]
+        'where' => [
+            'ads.recno' => $addId
+        ]
     ]);
 
-   $listing = $listingService->single([
-    'where' => [
-        'listingkeys' => $addInfo['listingcode']
-    ]
-   ]);
+    $listing = $listingService->single([
+        'where' => [
+            'listingkeys' => $addInfo['listingcode']
+        ]
+       ]);
+
+
+    if($addInfo['star_id']) {
+        Flash::set("Unable to delete ads with stars, please remove star from ads then delete again.", 'danger');
+        return redirect(_route('prop_show', [
+            'recno' => seal($listing['recno'])
+        ]));
+    }
 
    $adService->delete([
     'recno' => $addId
